pyproject.toml
[tool.poetry]
name = "tshirt-fulfillment"
version = "0.1.0"
description = "T-shirt Fulfillment AI Agent"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "tshirt_fulfillment", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
sqlalchemy = "^2.0.25"
pydantic = "^2.6.0"
redis = "^5.0.1"
google-auth = "^2.27.0"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.118.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.2.1"
pyright = "^1.1.350"
pre-commit = "^3.6.0"
pytest = "^8.0.0"
black = "^24.1.1"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "B008",  # Do not perform function call in argument defaults
]
fixable = ["ALL"]

[tool.ruff.lint.isort]
known-first-party = ["tshirt_fulfillment"]
combine-as-imports = true
force-wrap-aliases = true
force-single-line = true

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ALL"]
"*/tests/*" = ["ALL"]
"test_*.py" = ["ALL"]

[tool.pyright]
include = ["tshirt_fulfillment"]
exclude = ["**/node_modules", "**/__pycache__"]
pythonVersion = "3.9"
pythonPlatform = "All"
typeCheckingMode = "basic"
